+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 
+- `GradleRunner.gradlew` function
 - value argument to set the resulting task to `testBuildCache` and `testConfigurationCache`
 
 ### Changed